requirements.txt
cvzone==1.5.6
ultralytics==8.0.26
hydra-core>=1.2.0
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python==4.5.4.60
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.64.0
filterpy==1.4.5
scikit-image==0.19.3
lap==0.4.0
pygame>=1.9.6
customtkinter>=1.8